Release V1.8.3-alpha

+++ app/Http/Controllers/ActivityTypeController.php
@@ -96,6 +96,11 @@ class ActivityTypeController extends AppBaseController
      */
     public function destroy(ActivityType $activityType)
     {
+        if ($activityType->timeEntries()->count() > 0) {
+            return $this->sendError('This activity has more than one time entry, so it can\'t be deleted.');
+        }
+
+        $activityType->update(['deleted_by' => getLoggedInUserId()]);
         $activityType->delete();
 
         return $this->sendSuccess('Activity Type deleted successfully.');
